refactor(agent): split Agent Work Lab domains
